afs: Zero settime deltas correctly
Currently we zero the 'deltas' pointer itself, not the memory we
allocated for it. That's clearly wrong, and should cause a crash
whenever we access 'deltas'. Zero the pointed-to memory instead.

Thanks to Ryan Underwood who pointed out this issue.

This is a 1.6-only commit. On master, the settime code has been
removed entirely.

Change-Id: I0fdea89fb23945246dcbe83b18ec79be0f725fde
Reviewed-on: https://gerrit.openafs.org/11706
Reviewed-by: Perry Ruiter <pruiter@sinenomine.net>
Reviewed-by: Jeffrey Altman <jaltman@auristor.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>